Minor code beautify
